Site fonctionnel

+++ projetJeeSpringBoot/src/main/java/com/mainapp/controller/ManageFidelityPointController.java
@@ -41,11 +41,11 @@ public class ManageFidelityPointController {
 		if (fidelityPointList != null && customerList != null) {
 			Customer customer = null;
 			String email = null;
-			int fidelityPoint = 0;
+			double fidelityPoint = 0;
 			//Update the fidelity points of each customer
     		for (int i = 0; i < customerList.length; i++) {
     			email = customerList[i];
-    			fidelityPoint = Integer.parseInt(fidelityPointList[i]);
+    			fidelityPoint = Double.parseDouble(fidelityPointList[i]);
     			customer = customerService.getCustomer(email);
     			if (!customerService.setFidelityPoint(customer, fidelityPoint - customer.getFidelityPoint())) {
     				model.addAttribute("showAlert", "<script>showAlert('Update failed.', 'error', './ManageFidelityPoint')</script>");
